# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Author:
    properties:
      books:
        items:
          $ref: '#/definitions/models.Book'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  models.Book:
    properties:
      Updated_at:
        type: string
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      authors_id:
        items:
          type: integer
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      categories_id:
        items:
          type: integer
        type: array
      create_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      image:
        type: string
      resume:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      title:
        type: string
    type: object
  models.Category:
    properties:
      books:
        items:
          $ref: '#/definitions/models.Book'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  models.Error:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
...
  models.Review:
    properties:
      Updated_at:
        type: string
      book:
        $ref: '#/definitions/models.Book'
      book_id:
        type: integer
      comment:
        type: string
      create_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      score:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.User:
    properties:
      Updated_at:
        type: string
      access_token:
        type: string
      create_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        minLength: 6
        type: string
      password:
        type: string
      photo:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    required:
    - email
    - name
    - password
    - photo
    type: object
info:
  contact: {}
paths:
...
  /crud/review:
    get:
      description: Get all reviews
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Review'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get all reviews
      tags:
      - Review
    post:
      description: Create review
      parameters:
      - description: Review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create review
      tags:
      - Review
  /crud/review/{id}:
    delete:
      description: Delete review
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete review
      tags:
      - Review
    get:
      description: Get review by id
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get review by id
      tags:
      - Review
  /crud/review/book/{id}:
    get:
      description: Get all reviews by book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Review'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get all reviews by book
      tags:
      - Review
  /crud/user:
    get:
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.User'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get all users
      tags:
      - User
  /crud/user/{id}:
    delete:
      description: Delete user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete user
      tags:
      - User
    get:
      description: Get user by id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get user by id
      tags:
      - User
...
swagger: "2.0"